+++ tests/functional/errors/metadata-polyfill.ts
@@ -1,10 +1,16 @@
 import { ReflectMetadataMissingError } from "type-graphql";
-import { getMetadataStorage } from "@/metadata/getMetadataStorage";
+import { findType } from "@/helpers/findType";
 import { expectToThrow } from "../../helpers/expectToThrow";
 
 describe("Reflect metadata", () => {
   it("should throw ReflectMetadataMissingError when no polyfill provided", async () => {
-    const error = await expectToThrow(() => getMetadataStorage());
+    const error = await expectToThrow(() =>
+      findType({
+        metadataKey: "design:type",
+        prototype: {},
+        propertyKey: "test",
+      }),
+    );
 
     expect(error).toBeInstanceOf(ReflectMetadataMissingError);
     expect(error.message).toContain("metadata");