+++ tests/helpers/expectToThrow.ts
@@ -5,5 +5,5 @@ export async function expectToThrow<TError extends Error>(call: () => unknown): 
     return error as TError;
   }
 
-  throw new Error("Unreachable");
+  throw new Error("You've expected a function to throw, but it didn't throw anything.");
 }